reorganize tests and tweak conditional structure for empty objects
