+++ lib/parser.js
@@ -28,38 +28,37 @@ function parseBooleanValue(input){
 function parseObjectValue(input){
 
     var obj = {};
-    var parserRegex = /([^\00\01\02\b]+)\00/y;
 
     //should have properties, with type and value
     do {
         var type = input.str[input.i];
        
-        //handle non-empty objects
-        if (type !== constants.special.objectEnd){
-            var m = execRegexAt(input.str, parserRegex, ++input.i);
-            var propName = m[1];
-
-            input.i += 1 + propName.length;
-            try {
-                var val;
-                switch (type) {
-                    case constants.types.object:
-                        val = parseObjectValue(input);
-                        break;
-                    case constants.types.string:
-                        val = parseStringValue(input);
-                        break;
-                    case constants.types.boolean:
-                        val = parseBooleanValue(input);
-                        break;
-                }
-                obj[propName] = val;
-            } catch (err) {
-                console.error('encountered error while handling property:', propName, err);
-            }
-        }
-        else
+        //handle empty objects
+        if (type === constants.special.objectEnd) {
             break;
+        }
+
+        var m = execRegexAt(input.str, /([^\00\01\02\b]+)\00/y, ++input.i);
+        var propName = m[1];
+
+        input.i += 1 + propName.length;
+        try {
+            var val;
+            switch (type) {
+                case constants.types.object:
+                    val = parseObjectValue(input);
+                    break;
+                case constants.types.string:
+                    val = parseStringValue(input);
+                    break;
+                case constants.types.boolean:
+                    val = parseBooleanValue(input);
+                    break;
+            }
+            obj[propName] = val;
+        } catch (err) {
+            console.error('encountered error while handling property:', propName, err);
+        }
 
     } while(input.i < input.str.length);
