+++ test/static/shortcuts1.vdf
