+++ test/static/shortcuts2.vdf
