+++ test/static/shortcuts3.vdf
