+++ test/write-read-tests.js
@@ -1,0 +1,42 @@
+var shortcut = require('../lib');
+var should = require('should');
+
+
+describe('write-read tests', function(){
+
+    it('writes and reads same object', function(done){
+
+        var input = {
+            a: 'one',
+            b: {
+                c: 'two'
+            },
+            d: true,
+            e: [
+                'hello',
+                'world'
+            ],
+            f: {
+                g: [
+                    {
+                        name: 'hello',
+                        value: 'hi'
+                    },
+                    {
+                        n: 'hola',
+                        v: 'ciao'
+                    }
+                ]
+            }
+        };
+
+        var str = shortcut.stringify(input);
+        var output = shortcut.parse(str);
+
+        should.deepEqual(input, output);
+        done();
+    });
+
+
+});
+
